chore: Rename bundle driver to hub
For consistency, use "hub" as the storage driver name.
